+++ NEWS.md
@@ -1,5 +1,8 @@
 # pandoc-ling 1.4
 
+## bugs
+
+- fixed bug with not-appearing header of interlinear in HTML output
 
 # pandoc-ling 1.3